{"schema":"decodex/commit/1","summary":"Split work journal types","authority":"manual"}
