+++ packages/elf-service/src/work_journal/service.rs
@@ -9,10 +9,13 @@ use crate::{
 	search,
 	work_journal::{
 		types::{
-			DEFAULT_SESSION_READBACK_LIMIT, ELF_WORK_JOURNAL_SCHEMA_V1, MAX_SESSION_READBACK_LIMIT,
-			MAX_STORAGE_SCAN_ROWS, WorkJournalEntryCreateRequest, WorkJournalEntryCreateResponse,
-			WorkJournalEntryFamily, WorkJournalEntryGetRequest, WorkJournalEntryResponse,
+			WorkJournalEntryCreateRequest, WorkJournalEntryCreateResponse, WorkJournalEntryFamily,
+			WorkJournalEntryGetRequest, WorkJournalEntryResponse,
 			WorkJournalSessionReadbackRequest, WorkJournalSessionReadbackResponse,
+			constants::{
+				DEFAULT_SESSION_READBACK_LIMIT, ELF_WORK_JOURNAL_SCHEMA_V1,
+				MAX_SESSION_READBACK_LIMIT, MAX_STORAGE_SCAN_ROWS,
+			},
 		},
 		validation::{self},
 	},
